[MOJO-1640] no need to copy target/classes of a dependency
git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/maven-native@13524 52ab4f32-60fc-0310-b215-8acea882cd1b

+++ native-maven-plugin/src/main/java/org/codehaus/mojo/natives/plugin/NativeLinkMojo.java
@@ -446,7 +446,7 @@ public class NativeLinkMojo
 
         try
         {
-            if ( doCopy && ( !newLocation.exists() || newLocation.lastModified() <= artifact.getFile().lastModified() ) )
+            if ( doCopy && !artifact.getFile().isDirectory() && ( !newLocation.exists() || newLocation.lastModified() <= artifact.getFile().lastModified() ) )
             {
                 FileUtils.copyFile( artifact.getFile(), newLocation );
             }
@@ -463,7 +463,7 @@ public class NativeLinkMojo
     ////////////////////////////////////// UNIT TEST HELPERS //////////////////////////////////
 
     /**
-     * For unittest only
+     * For unit test only
      */
     private LinkerConfiguration config;
